Fix: Class "Account" has 21 methods, which is greater than 20 authorized. Split it into smaller classes.
